Generowanie lektorów - rand to jest jakiś żart :(

+++ LangSystem_Generator/Models/Database/Lector.cs
@@ -8,13 +8,13 @@ namespace LangSystem_Generator.Models.Database
 {
     class Lector
     {
-        public int PESEL { get; set; }
+        public long PESEL { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string Department {get; set;}
         public string Language {get; set;}
 
-        public Lector(int pesel, string firstName, string lastName, string department, string language)
+        public Lector(long pesel, string firstName, string lastName, string department, string language)
         {
             this.PESEL = pesel;
             this.FirstName = firstName;
